Moved parametric generate_lnu
